refactor : 디렉토리 구조 DDD style로 변경
