+++ src/app.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthModule } from './auth/auth.module';
+import { CommonModule } from './common/common.module';
 import { typeORMConfig } from './common/config/typeorm.config';
 import { UserModule } from './user/user.module';
 import { WorkbookModule } from './workbook/workbook.module';
@@ -11,6 +12,7 @@ import { WorkbookModule } from './workbook/workbook.module';
     AuthModule,
     WorkbookModule,
     UserModule,
+    CommonModule,
   ],
 })
 export class AppModule {}
