--- src/auth/decorators/get-user.decorator.ts
@@ -1,9 +1,0 @@
-import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { User } from '../entities/user.entity';
-
-export const GetUser = createParamDecorator(
-  (data, ctx: ExecutionContext): User => {
-    const request = ctx.switchToHttp().getRequest();
-    return request.user;
-  },
-);
