+++ src/auth/domain/user.entity.ts
