+++ src/auth/infrastructure/user.repository.ts
@@ -1,8 +1,8 @@
 import { BcryptUtils } from '@/common/utils/bcrypt.utils';
 import { ConflictException } from '@nestjs/common';
 import { EntityRepository, Repository } from 'typeorm';
-import { SignUpRequest } from './dto/sign-up.request';
-import { User } from './entities/user.entity';
+import { User } from '../domain/user.entity';
+import { SignUpRequest } from '../interfaces/sign-up.request';
 @EntityRepository(User)
 export class UserRepository extends Repository<User> {
   async signUp(signUpRequest: SignUpRequest): Promise<User> {
