+++ src/user/test/interfaces/user.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { UserController } from './user.controller';
-import { UserService } from './user.service';
+import { UserService } from '../../application/user.service';
+import { UserController } from '../../interfaces/user.controller';
 
 describe('UserController', () => {
   let controller: UserController;
