+++ src/workbook/domain/workbook-card.entity.ts
