+++ src/workbook/interfaces/update-workbook-like-request.ts
@@ -1,5 +1,5 @@
 import { IsEnum } from 'class-validator';
-import { WorkbookLikeType } from '../entities/workbook-like.entity';
+import { WorkbookLikeType } from '../domain/workbook-like.entity';
 
 export class UpdateWorkbookLikeRequest {
   @IsEnum(WorkbookLikeType)
