+++ src/workbook/interfaces/workbook.response.ts
@@ -1,5 +1,5 @@
-import { UserInfoResponse } from '@/user/dto/user-info.response';
-import { Workbook } from '../entities/workbook.entity';
+import { UserInfoResponse } from '@/user/interfaces/user-info.response';
+import { Workbook } from '../domain/workbook.entity';
 
 export class WorkbookResponse {
   id: number;
